requirements.txt
requests
pillow
scikit-image
fastai==1.0.61
numpy==1.21.5
pandas==1.1.5 
torch==1.12.0
torchvision==0.13.0
opencv-python==4.5.4.58
timm==0.6.11
ipython==8.4.0
blend-modes==2.1.0
